@@ -1,6 +1,6 @@
-const { getInstallationOctokit } = require('../config/github-app');
-const aiService = require('../services/ai-service');
-const logger = require('../utils/logger');
+import { getInstallationOctokit } from '../config/github-app.js';
+import aiService from '../services/ai-service.js';
+import logger from '../utils/logger.js';
 
 /**
  * Handler para quando um PR é aberto
@@ -295,7 +295,7 @@ function calculateTimeOpen(createdAt, closedAt) {
   }
 }
 
-module.exports = {
+export {
   handlePullRequestOpened,
   handlePullRequestEdited,
   handlePullRequestClosed,
